AqXco 0.58
Modificado:
-colisionComponent
-cargadorComponent
-system
-Entity2D
facadeColision

+++ DESARROLLO/HITO1/ITERACION 3/ArquitecturaXComponentes/entitySystem/systems/system.cpp
@@ -29,7 +29,7 @@ system::~system(){
     std::cout<<"Sistema eleminado"<<std::endl;
 }
 
-void system::update(unsigned int dt){
+void system::update(float dt){
 }
 
 entityManager* system::getEntityManager(){
